requirements.txt
Flask==2.3.1
Pillow==9.5.0
selenium==4.9.0